+++ carp_core/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.1.0 - BREAKING
 
 * the first stable release of CARP cores which follows the [carp.core-kotlin](https://github.com/imotions/carp.core-kotlin) v. 1.1.0 domain model.
-* upgrade to [carp_serializable](https://pub.dev/packages/carp_serializable) v. 1.1.0
+* upgrade to [carp_serializable](https://pub.dev/packages/carp_serializable) v. 1.1.0.
+* added `PersonalComputer` and `WebBrowser` as new device configurations.
 
 ## 0.40.8